allowing setting image in runtime with a file

+++ test_programs/pwm_two_servos/svg/hello_world.h
@@ -3,5 +3,5 @@
 
 #include <vector>
 
-extern std::vector<std::vector<std::pair<double, double>>> hello_world;
+extern std::vector<std::vector<std::pair<float, float>>> hello_world;
 #endif // HELLO_WORLD_H
